Rename launchable module to nose_launchable module
