--- launchable/__init__.py
@@ -1,5 +1,0 @@
-import sys
-
-from launchable.plugin import Launchable  # noqa: F401
-
-sys.setrecursionlimit(10000)
