+++ nose_launchable/__init__.py
@@ -1,0 +1,5 @@
+import sys
+
+from nose_launchable.plugin import Launchable  # noqa: F401
+
+sys.setrecursionlimit(10000)
